requirements/base.txt
pytz==2022.1  # https://github.com/stub42/pytz
python-slugify==6.1.2  # https://github.com/un33k/python-slugify
Pillow==9.2.0  # https://github.com/python-pillow/Pillow
rcssmin==1.1.0  # https://github.com/ndparker/rcssmin
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.2.0  # https://github.com/evansd/whitenoise
redis==4.3.4  # https://github.com/redis/redis-py

# Django
# ------------------------------------------------------------------------------
django==4.0.5  # pyup: < 4.0  # https://www.djangoproject.com/
django-environ==0.9.0  # https://github.com/joke2k/django-environ
django-model-utils==4.2.0  # https://github.com/jazzband/django-model-utils
# django-allauth==0.51.0  # https://github.com/pennersr/django-allauth
django-compressor==4.0  # https://github.com/django-compressor/django-compressor
django-redis==5.2.0  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.13.1  # https://github.com/encode/django-rest-framework
django-cors-headers==3.13.0 # https://github.com/adamchainz/django-cors-headers
# DRF-spectacular for api documentation
drf-spectacular==0.22.1  # https://github.com/tfranzel/drf-spectacular

# Custom
# ------------------------------------------------------------------------------
django-jazzmin==2.5.0  # https://github.com/farridav/django-jazzmin
dj-rest-auth[with_social]==2.2.4
django_filter==22.1
